* 断货天数需要展示小数

+++ core/app/models/view/dto/AnalyzeDTO.java
@@ -137,7 +137,7 @@ public class AnalyzeDTO implements Serializable {
     /**
      * 断货天数
      */
-    public int outday;
+    public float outday;
 
     /**
      * 实时价格
